Review fixes. However, somethings broken with TaskUpdateImportsTest
The test seems to encounter a theme import which is terminated in an extra
quote (").
